back-py/requirements.txt
fastapi
xrpl-py